refactor: 불필요한 코드 정리 및 오해 유발 이름 수정
- firebase.ts → google-auth.ts 이름 변경 (Firebase SDK 미사용)
- export type User 스텁 제거
- proxy.ts → middleware.ts 이동 (Next.js 미들웨어로 실제 동작)
- formatCredits 제거 (프로덕션 미사용)
- CREDIT_RATE_PER_MINUTE 중복 선언 제거 (plans.ts에서 삭제)
