+++ middleware.ts
@@ -2,7 +2,7 @@ import { NextResponse } from 'next/server'
 import type { NextRequest } from 'next/server'
 import { SESSION_COOKIE, verifySessionCookie } from '@/lib/auth/session-cookie'
 
-export function proxy(request: NextRequest) {
+export function middleware(request: NextRequest) {
   const raw = request.cookies.get(SESSION_COOKIE)?.value
   if (!raw || !verifySessionCookie(raw)) {
     return NextResponse.redirect(new URL('/', request.url))
