+++ src/utils/formatters.ts
@@ -6,10 +6,6 @@ export function formatDuration(seconds: number): string {
   return `${m}:${s.toString().padStart(2, '0')}`
 }
 
-export function formatCredits(credits: number): string {
-  return credits.toLocaleString()
-}
-
 export function formatCurrency(amount: number): string {
   return new Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD' }).format(amount)
 }
